Unify the Venus steering model on one firmware-verified law
The VNSD-0 images settle what the VNSE3-0 ones could only suggest: the Venus A,
D and E run *the same* CT-following law, so the simulator should have one model
for the family rather than the three different ones it had (VNSD-0 on a bare
integrator with no gate, VNSE3-0 and VNSA-0 on the HMG-50 float ramp).

Verified across both archives:

  * VNSD-0's integrator (0x0802c8c4 in v150) and VNSE3-0's (0x08029ecc) are the
    same code instruction for instruction, as are their gates (0x080266a0 /
    0x08023d2c) and their share splits (0x080042b8 / 0x08004218). The VNSA-0
    integrator matches too. The law is unchanged across VNSD-0
    v147/v149/v1492/v150 and VNSE3-0 v144/v148/v1476/v150.
  * None of the three contains any of the eleven HMG-50 float gain-table
    constants; all ten archived HMG-50 images contain all eleven.

So the Venus D model was wrong in three ways the Venus E model had already
fixed: it had no input gate at all, its per-step branch was fed the CT value
rather than the unit's own output, and its +-15 W park was selected by "phase D"
when the firmware selects it by the bucket's *_chrg_nb count -- which also
share-splits the reading.

venus_d_steering.py and venus_e_steering.py are replaced by
venus_integer_steering.py, which carries the law, the gate and the split
together (step_raw is the bare law, as in firmware_steering) and cites the
firmware addresses for each part in both archives. battery.py routes every
AC-coupled VNS* device to it.

Steering evaluation over the five Venus D scenarios, 3 seeds: 12 metrics
improved, 2 regressed, no guardrail regressions -- cost regret -15%, battery
travel -10%, settle p95 -12%. The Venus E scenarios are unchanged from the
previous commit, which already ran this law.
